requirements.txt
django==2.0.5
PyMySQL==0.8.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-cors-headers==2.2.0
mysqlclient==1.3.12